+++ src/DBInterface/DBInterface.cpp
@@ -9,6 +9,8 @@ void DBInterface::closeConnection() {
     connection.close();
 }
 
+/// Creates the necessary tables if they don't exist already, compiles the prepared Statements for later queries
+/// and inserts the event types into the eventtypes table (if they do not already exist)
 void DBInterface::setupDB() {
     pqxx::work transaction(connection);
     const std::string eventTypesTable = "CREATE TABLE IF NOT EXISTS eventtypes ("
@@ -34,6 +36,9 @@ void DBInterface::setupDB() {
     transaction.commit();
 }
 
+/// Inserts and event into the events table. You have to either provide a targetID
+/// (which can be acquired using getTargetIDWithAddress) or set a default target before
+/// calling this function using setDefaultAddress
 void DBInterface::insertEvent(const std::string& description, enum EventType eventType, int targetID) {
     int tID = (targetID==-1) ? defaultAddressID : targetID;
     assert(tID!=-1);
@@ -42,6 +47,7 @@ void DBInterface::insertEvent(const std::string& description, enum EventType eve
     transaction.commit();
 }
 
+/// Sets the defaultAddressID used by insertEvent if no targetID is provided.
 void DBInterface::setDefaultAddress(const std::string& address) {
     defaultAddressID=getTargetIDWithAddress(address);
 }
@@ -54,7 +60,7 @@ void DBInterface::prepareStatements() {
     connection.prepare("getTargetID",
                        "SELECT id FROM pingtargets where address=$1");
     connection.prepare("registerEvent",
-                       "INSERT INTO eventtypes VALUES($1)");
+                       "INSERT INTO eventtypes VALUES($1) ON CONFLICT DO NOTHING");
 }
 
 /// attempts to insert address into pingtargets.
@@ -72,18 +78,28 @@ bool DBInterface::insertTarget(const std::string& address) {
     }
 }
 
+/// retrieves the id of the provided address from the database.
+/// If the address is not in the database -1 is returned instead.
 int DBInterface::getTargetIDWithAddress(const std::string& address) {
     pqxx::work transaction(connection);
-    pqxx::row row = transaction.exec_prepared1("getTargetID", address);
-    transaction.commit();
-    return row[0].as<int>();
-//    TODO: Test for nonexistant address
+    try{
+        pqxx::row row = transaction.exec_prepared1("getTargetID", address);
+        transaction.commit();
+        return row[0].as<int>();
+    }
+    catch(pqxx::unexpected_rows& e){
+        transaction.commit();
+        return -1;
+    }
 }
 
 /// inserts the event strings specified in Events.h into the database
 void DBInterface::addEventTypes(pqxx::transaction_base& transaction) {
     for(const std::string& s:EventStrings){
-        transaction.exec_prepared("registerEvent",s);
+        try{
+            transaction.exec_prepared("registerEvent",s);
+        }
+        catch(pqxx::unique_violation& e){}
     }
 }
